+++ src/lib/custom-typings/glob-watcher.d.ts
@@ -1,0 +1,8 @@
+declare module "glob-watcher" {
+  namespace globWatcher {
+  }
+
+  const globWatcher: any;
+
+  export = globWatcher;
+}