+++ src/lib/index.ts
@@ -14,10 +14,5 @@ export {
 } from "./options/tslint";
 export { OutModules, TypescriptOptions } from "./options/typescript";
 export { DEFAULT_PROJECT, Project } from "./project";
-export {
-  Target,
-} from "./targets";
-export { generateLibTasks, LibTarget, registerLibTasks } from "./targets/lib";
-export { generateMochaTasks, MochaTarget, registerMochaTasks } from "./targets/mocha";
-export { generateNodeTasks, NodeTarget, registerNodeTasks } from "./targets/node";
+export { Target } from "./target";
 export { PackageJson } from "./utils/project";