+++ src/lib/project-tasks/bump-major.ts
@@ -1,11 +1,11 @@
-import { Gulp } from "gulp";
+import Undertaker from "undertaker";
 import { Project } from "../project";
 import { bumpVersion } from "../utils/project";
 
 export const taskName: string = ":bump-major";
 
-export function registerTask(gulp: Gulp, project: Project): void {
-  gulp.task(taskName, async function (): Promise<void> {
+export function registerTask(taker: Undertaker, project: Project): void {
+  taker.task(taskName, async function (): Promise<void> {
     return bumpVersion("major", project);
   });
 }