+++ src/lib/task-generators/copy.ts
@@ -1,7 +1,8 @@
 import { FSWatcher } from "fs";
-import { Gulp } from "gulp";
+import globWatcher from "glob-watcher";
 import { Minimatch } from "minimatch";
-import { TaskFunction } from "../utils/gulp-task-function";
+import { TaskFunction } from "undertaker";
+import vinylFs from "vinyl-fs";
 import { asString, join } from "../utils/matcher";
 
 export interface Options {
@@ -28,25 +29,25 @@ export function getSources({files, from}: Options): string[] {
   return files.map((val: string): string => asString(join(from, new Minimatch(val))));
 }
 
-export function copy(gulp: Gulp, options: Options): NodeJS.ReadableStream {
-  return gulp
+export function copy(options: Options): NodeJS.ReadableStream {
+  return vinylFs
     .src(getSources(options), {base: options.from})
-    .pipe(gulp.dest(options.to));
+    .pipe(vinylFs.dest(options.to));
 }
 
 /**
  * Generate a task to copy files from one directory to an other.
  */
-export function generateTask(gulp: Gulp, options: Options): TaskFunction {
+export function generateTask(options: Options): TaskFunction {
   const task: TaskFunction = function (): NodeJS.ReadableStream {
-    return copy(gulp, options);
+    return copy(options);
   };
   task.displayName = "_copy";
   return task;
 }
 
-export function watch(gulp: Gulp, options: Options): FSWatcher {
-  const buildTask: TaskFunction = generateTask(gulp, options);
+export function watch(options: Options): FSWatcher {
+  const buildTask: TaskFunction = generateTask(options);
   const sources: string[] = getSources(options);
-  return gulp.watch(sources, {cwd: options.from}, buildTask) as FSWatcher;
+  return globWatcher(sources, {cwd: options.from}, buildTask) as FSWatcher;
 }