+++ src/lib/utils/gulp-task-function.ts
@@ -1,3 +1,0 @@
-import { TaskFunc } from "orchestrator";
-
-export type TaskFunction = TaskFunc & {name?: string; displayName?: string; description?: string} | any;